// react-zustand/src/types/product.ts
export interface Product {
	id: number
	title: string
	description: string
	category: string
	price: number
	discountPercentage: number
	rating: number
	stock: number
	brand?: string
	thumbnail?: string
	images?: string[]
}

export interface ProductsResponse {
	products: Product[]
	total: number
	skip: number
	limit: number
}

export type ProductDraft = Omit<
	Product,
	'id' | 'rating' | 'thumbnail' | 'images'
> & {
	thumbnail?: string
}